+++ routes/api/signUp.js
@@ -1,47 +1,11 @@
 const express = require('express');
 const router = express.Router();
-const {signUpValidation} = require('../../validation');
-const bcrypt = require('bcryptjs');
 
-//User Model 
-const User = require('../../models/User');
+// Controller
+const {signUp} = require('./../../controllers/authentication_controller');
 
 // @route   POST api/signUp
 // @desc    Create User
-router.post('/', async (req,res) =>{
-
-    //Validating SignUp Body:
-    const {error} = signUpValidation(req.body)
-    if (error) return res.status(400).send(error);
-
-
-    //Checking is user is already in DB:
-    const emailExist = await User.findOne({ EmailAddress: req.body.EmailAddress });
-    if (emailExist != null) return res.status(400).send({ error: 'Email already exists' });
-
-    //Hash Passwords:
-    const salt = await bcrypt.genSalt(10);
-    const hashPassword = await bcrypt.hash(req.body.Password,salt);
-    
-    const post = new User({
-        FirstName: req.body.FirstName,
-        LastName: req.body.LastName,
-        Suburb: req.body.Suburb,
-        Password: hashPassword,
-        EmailAddress: req.body.EmailAddress,
-        Age: req.body.Age
-    });
-  
-    try {
-        const savedPost = await post.save();
-        res.json({ 
-                user: savedPost._id 
-            });
-    } catch(err){
-        res.json({
-            message:err
-        });
-    }
-});
+router.post('/', signUp);
 
 module.exports = router;
